Move service_offering_check to a mixin
